pyproject.toml
[tool.poetry]
name = "dowado-web-server"
version = "0.1.0"
description = ""
authors = ["taehwan heo <htth815@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
langchain = "^0.2.7"
pydantic = "^2.8.2"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.16"
langchain-core = "^0.2.18"
redis = "^5.0.7"
dacite = "^1.8.1"
langchain-postgres = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pycryptodome = "^3.20.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pillow = "^10.4.0"
bcrypt = "^4.1.3"
torch = "^2.3.1"
kiwipiepy = "^0.18.0"
transformers = "^4.43.1"
pyjwt = "^2.8.0"
huggingface-hub = "^0.24.6"
greenlet = "^3.0.3"
langchain-huggingface = "^0.0.3"
langchain-teddynote = "^0.0.30"
vllm = "^0.6.0"
accelerate = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
